requirements.txt
asgiref==3.8.1
coverage==7.8.0
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
